Enforcing style guide
Trying to standardize language in prompts; specifically component names and category names are title-cased, messages and
subcomponent are cased as sentences. Also tried to standardize terminology for terms like 'multi-class' and 'PnP'. Author
names now appear in brackets, not parentheses. Comments about game applicability are removed as the installer should handle it.