Romance Accelerator
As noted by Lord Tansheron, the Romance Accelerator is a bit spotty with the EE NPCs. Mainly, it was due to some timers being
set by the dialogue files, not scripts, and some using a non-GLOBAL variable scope so that the match would fail. Since the
framework was built, it's been extended to the pseudo-romances for the EE NPCs in BGEE as well. SoD romances are event-driven
so they're still ignored for this component.