Means testing for components with item description updates
For components that update item descriptions, we generally just checked if the description string was in a valid range of 0 to
2147483646, ran some text replacement macros, and then SAY_EVALUATE'd it back onto the item. However, if the string was actually
invalid--say it was assigned strref #60000 on a 50,000 string tlk table--this practice would result in reading in the contents as
WeiDU's default error content of "<Invalid Strref 60000>" and then placed on the item as a valid string.

Instead of screening on a fixed numeric range for strrefs, we now read the value directly and instead proceed only if the strref
contents are not the default error message. This change affects multiple components: the 2H weapons, all of the proficiency
changes, universal clubs, thieving and arcane casting in heavy armor, universal clubs, zero-weight magical items, consistent
+x/+y items, and PnP protection items.