Fixes for Send BioWare NPCs to an Inn
Rasaad and Neera have actions in their script to EscapeAreaMove if they're not in the party, which this component disables via
REPLACE_TEXTUALLY. However, if More Interjections is istalled, the InParty triggers can decompile as IVFPD (depending on the
order of trigger.ids) causing these patches to fail and laving the scripting intact.

Neera's post-dialogue is also patched to remove her default move back to the FAI. However, because it was being done after the
the new options were added, the new action to move to FAI was also being cleared.