Remove XP cap
A semi-intelligent* attempt to extend kit ability tables from level 40 to level 50. Originally Tweaks copied levels 31-40 and
appended them column-wise to create entries for 41-50. In the last few versions it was swapped over to simply build out the
table with blank entries since the EEs shipped with some malformed tables. It now starts with that, and then attempts to
backfill by checking for known 'per X level' abilities. In the worst case, it will copy from 12 levels prior since that should
catch the most per-level iterations (1/2/3/4/6/12).

* Semi-intelligent, in this case, refers to the algorithm. The author remains a non-literate monkey banging at his keyboard.