Fixing issue with Friendly Random Drops/IWDFP conflict
oIWD has duplicate random drops (extres4) in ar9200 and ar9601. IWDEE and older versions of IWDFP (v6 or older) replaced the
ar9200 item with extres1. As of v7, IWDFP instead replaces the ar9601 dupe with extres1.

Friendly Random Drops was updated to detect IWDFP >= v7 and account for this... except it did it half-assed and could apply the
old fix itself. Another half-ass has been added, and we should now be fully-assed.