PnP Proficiency for Dual-Classes Grants F/M 1 Extra Spell Per Level
https://www.gibberlings3.net/forums/topic/36465-bug-pnp-proficiency-for-dual-classes-grants-fm-1-extra-spell-per-level/

jmerry diagnoses and provides the fix for EEs, which is implemented. For non-EEs, things get a little tricky since the trigger
that fixes EEs doesn't exist. So, we settle on a racial check since only humans can dual-class.

Unfortunately, there's a really bad mod with a terrible author that has options to allow humans to multi-class and/or to allow
non-humans to dual-class on the originals with TobEx. Until a clever solution is found for this edge case, this component
will get skipped on such games.