Added new component: Sensible Regeneration When Traveling
When traveling on the worldmap it was always frustrating that, after a 20-hour transit, only the character wearing a Ring of
Regeneration gets healed. A sensible party would pass the ring around to get everyone healed on such a long trip, the way a
player might swap items when walking across a map.

This component will search for regeneration items above a certain threshold--currently set at 1 hp/9 seconds, enough to include
Rings of Regeneration--and, if the party is in possession, simply heals the party fully after any worldmap travel that isn't
instant. It's certainly not precise, as items which aren't usable by everyone get counted as regen items (e.g. Axe of
Unyielding), but that's why it's in the convenience/cheats section.