Using Argent's batch scheme
Argent offered an alternative to the current batch install scheme of fake installations by leveraging ABORT. This allows the
installer to move ahead to the next component without the need to fake it; prior to the ABORT we register the install request
into a batch script which runs on exit. This approach allows batch mode to be invoked on every install of Tweaks, not just a clean
fist-time install. More importantly it also allows for safer failure conditions, namely that components are either installed or
not, with no ambiguity due to faking component installations.

In addition to the change in the baseline approach, he also had a much more thorough vetting of the invoking executable, which
I've also adopted.