Updates to Joinable vs. Non-Joinable creature handling and Max HP NPCs
First, we're using an adaptation of k4thos's joinable NPC detection used in EET (see joinable_npc_array.tpa) to build an array
of creature files of joinable vs. non-joinable NPCs. As part of this we had to expand the array to cover non-EE games: oBG and
oBG2 were fine as they use the same mechanism (pdialog) for detection as the EEs. PsT and IWD2 (if using the NPCs from IWD2EE)
have manual lists. The only oIWD NPCs are CHR files so nothing gets filed as joinable.

As a side effect of this, the Max HP Creature components has been expanded to cover 'proper' NPCs in IWD2EE, and will now patch
character files on all platforms save PsT/PsTEE (as it doesn't have any). The character files are considered "joinable"
creatures. In this way the NPCs from IWD NPC (when in oIWD) and IWD2 NPC also get covered. As oIWD and IWD2 have "joinable"
NPCs as a result of this change, all variants of the component are available to all games.

Max HP also now accounts for non-standard die kits in the original game, namely, that Dwarven Defenders roll d12 for HP and
Dragon Disciples roll d6. While it remains possible to do this down to the kit and level for every modded kit and class, I
simply don't have the energy to do all of the lookups required.