Compat fix between 2H weapon components and altered prof components
The various 2H weapon components were expected to be installed prior to the various components that changed proficiency
schemes. As a result there were various assumptions that eventually ended up with an install error.

These components will now properly detect and adjust as needed.

Credit to Connelly for the bug report.